Fix auto-commands eating letters from operator names
Should have lower "precedence".

Auto-commands used to be the first thing that Letter::createLeftOf()
checked for. Now instead it first waits for the dust to settle on
auto-operator names, then seeks nearby letters but stops at the edge of
an operator name. Also note that before, this letter itself had not been
inserted yet; now it has so the seeking and the fragment that's removed
both start from the this letter.